# Replace indexing with `get_unchecked` calls